Add getGenres to manhwa service
