+++ manhwa-nexus-web-client/package.json
@@ -21,6 +21,7 @@
         "@angular/platform-browser": "~13.2.0",
         "@angular/platform-browser-dynamic": "~13.2.0",
         "@angular/router": "~13.2.0",
+        "@ng-select/ng-select": "^10.0.3",
         "rxjs": "~7.5.0",
         "tslib": "^2.3.0",
         "zone.js": "~0.11.4"
