+++ manhwa-nexus-web-client/src/app/pages/browse/browse.component.html
@@ -1,1 +1,3 @@
-<p>browse works!</p>
+<div class="page-container">
+    <div class="filters"></div>
+</div>
