+++ manhwa-nexus-web-client/src/app/services/config.service.spec.ts
@@ -1,6 +1,6 @@
 import { TestBed } from '@angular/core/testing';
 
-import { ConfigService } from './config-service.service';
+import { ConfigService } from './config.service';
 
 describe('ConfigServiceService', () => {
     let service: ConfigService;
