+++ manhwa-nexus-web-client/src/app/services/config.service.ts
