+++ manhwa-nexus-web-client/src/assets/client-config.json
@@ -1,6 +1,6 @@
 {
-  "api": {
-    "server": "http://localhost:8080/api/v1/",
-    "gateway": "http://localhost:8700/"
-  }
+    "api": {
+        "server": "http://localhost:8080/api/v1",
+        "gateway": "http://localhost:8700/"
+    }
 }
